requirements.txt
wheel
qsimcirq
fasteners
pyzx
pytest
psutil